+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - add mock books to dev page
 - add stats to dev page
 - add more translations
+- implement more provider handling (AddNewbookPage)
 
 ## [0.8.1] - 2025-11-20
 - library sorting to case-insensitive